Change the kernel mmap interface so that the offset to map is an "off_t" and the return value is a "paddr_t" to allow mappings at offsets past 2^31 bytes.  Somewhat inspired by FreeBSD, which only changed the offset to a "vm_offset_t".
Includes updates for the i386, pc532 and sh3 mmmmap from Jason Thorpe.

+++ sys/arch/hp300/dev/grfvar.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: grfvar.h,v 1.11 1998/06/25 23:57:34 thorpej Exp $	*/
+/*	$NetBSD: grfvar.h,v 1.12 2000/06/26 04:55:39 simonb Exp $	*/
 
 /*
  * Copyright (c) 1988 University of Utah.
@@ -129,7 +129,7 @@ int	grfmap __P((dev_t, caddr_t *, struct proc *));
 int	grfunmap __P((dev_t, caddr_t, struct proc *));
 int	grfon __P((dev_t));
 int	grfoff __P((dev_t));
-int	grfaddr __P((struct grf_softc *, int));
+off_t	grfaddr __P((struct grf_softc *, off_t));
 
 #ifndef _LKM
 #include "opt_compat_hpux.h"
